Made Health in Entity floating point, moved a function call to constructor in game.cpp, and renamed User usr to user

+++ Engine/Entity.cpp
@@ -1,6 +1,6 @@
 #include "Entity.h"
 
-void Entity::TakeDamage(int DamageValue)
+void Entity::TakeDamage(float DamageValue)
 {
 	Health = Health - DamageValue;
 };
